requirements.txt
Django==4.2.20
asgiref==3.8.1
sqlparse==0.5.3
typing-extensions==4.13.2
backports.zoneinfo==0.2.1
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
psycopg2-binary==2.9.10

